+++ LoginLdap.php
@@ -46,6 +46,8 @@ class LoginLdap extends \Piwik\Plugin
             'Controller.Login.confirmPassword'       => 'checkIfPasswordConfirmationCanBeSkipped',
             'UsersManager.checkPassword'             => 'checkPassword',
             'UsersManager.deleteUser'                => 'onUserDeleted',
+            'API.UsersManager.setUserPreference'     => 'checkIfUserPreferenceChangeIsAllowed',
+            'API.UsersManager.updateUser'            => 'disablePasswordUpdate',
             'Login.userRequiresPasswordConfirmation' => 'skipPasswordConfirmation',
         );
         return $hooks;
@@ -292,6 +294,33 @@ class LoginLdap extends \Piwik\Plugin
         if ($this->isUserLdapUser($userLogin)) {
             $key = $userLogin . UsersManagerAPI::OPTION_NAME_PREFERENCE_SEPARATOR . UserMapper::USER_PREFERENCE_NAME_IS_LDAP_USER;
             Option::delete($key);
+        }
+    }
+
+    /**
+     * Throws exception if trying to change isLDAPUser preference, which is not allowed via API
+     *
+     * @param $preferenceInfo
+     * @return void
+     * @throws Exception
+     */
+    public function checkIfUserPreferenceChangeIsAllowed($preferenceInfo)
+    {
+        if (isset($preferenceInfo['preferenceName']) && $preferenceInfo['preferenceName'] === UserMapper::USER_PREFERENCE_NAME_IS_LDAP_USER && !UserMapper::$isLdapPreferenceSetAllowed) {
+            throw new Exception(Piwik::translate('LoginLdap_LdapUserCantChangeIsLdapUserPreference'));
+        }
+    }
+
+    /**
+     * Throws exception if trying to change password of LDAP user
+     * @param $values
+     * @return void
+     * @throws Exception
+     */
+    public function disablePasswordUpdate($values)
+    {
+        if (!UserSynchronizer::$allowUpdateUser && !empty($values['userLogin']) && !empty($values['password']) && $this->isUserLdapUser($values['userLogin'])) {
+            throw new Exception(Piwik::translate('LoginLdap_LdapUserCantChangePassword'));
         }
     }